Moved timer into time module
